examples/usecases/codeforces_question_solver/requirements.txt
streamlit>=1.32.0
camel-ai[all]==0.2.59
python-dotenv>=1.0.0 
firecrawl-py==2.5.3
